Move utility functions to utils.h
